+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - PPDB files report `has_sources() = true` even when they only contain source links ([#774](https://github.com/getsentry/symbolic/pull/774))
 - Make SourceBundle writer deterministic ([#778](https://github.com/getsentry/symbolic/pull/778))
+- Increase IL2CPPP source mapping boundary ([#776](https://github.com/getsentry/symbolic/pull/776))
 
 ## 12.1.1